quiz_generator_utils 생성
